+++ packages/example/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.6](https://github.com/Portkey-Wallet/portkey-web/compare/v2.6.6-alpha.0...v2.6.6) (2024-05-30)
+
+**Note:** Version bump only for package @portkey/example
+
+## [2.6.6-alpha.0](https://github.com/Portkey-Wallet/portkey-web/compare/v2.6.5...v2.6.6-alpha.0) (2024-05-28)
+
+**Note:** Version bump only for package @portkey/example
+
 ## [2.6.5](https://github.com/Portkey-Wallet/portkey-web/compare/v2.6.4...v2.6.5) (2024-05-16)
 
 **Note:** Version bump only for package @portkey/example